backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.1.8
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
kafka-python==2.1.3
multidict==6.2.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.1.0
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
